refactoring selectedVideo relatedVideo
