+++ src/components/Header/Header.module.scss
@@ -5,6 +5,7 @@
   display: flex;
   height: 56px;
   background: #202020;
+  z-index: 99;
   width: 100%;
 
   a {
