+++ src/components/SideList/SideList.module.scss
@@ -2,4 +2,8 @@
   width: 480px;
   flex-basis: 480px;
   padding: 1rem;
+  @media screen and (max-width: 960px) {
+    width: 90%;
+    margin: 1rem auto;
+  }
 }
