+++ src/components/SideListItem/SideListItem.module.scss
@@ -4,12 +4,19 @@
   margin-bottom: 1rem;
   img {
     width: 50%;
+    @media screen and (max-width: 560px) {
+      width: 100%;
+    }
   }
   .info {
     display: inline-block;
     width: 46%;
     padding: 2%;
     vertical-align: top;
+    @media screen and (max-width: 560px) {
+      width: 100%;
+      padding: 0;
+    }
   }
   span {
     color: white;
