+++ src/components/layout/Layout.module.scss
@@ -1,5 +1,4 @@
 .wrapper {
-  height: 100vh;
   min-height: 100vh;
   display: flex;
   flex-direction: column;
@@ -10,5 +9,10 @@
     flex: 1;
     width: 100%;
     background: #181818;
+    @media screen and (max-width: 960px) {
+      max-width: 90%;
+      margin: 0 auto;
+      flex-direction: column;
+    }
   }
 }
