+++ src/pages/Top.js
@@ -9,12 +9,12 @@ const Top = () => {
   const { globalState, setGlobalState } = useContext(Store);
 
   useEffect(() => {
-    fetchPopularData().then((res) => {
-      setGlobalState({
-        type: "SET_POPULAR",
-        payload: { popular: res.data.items },
-      });
-    });
+    // fetchPopularData().then((res) => {
+    //   setGlobalState({
+    //     type: "SET_POPULAR",
+    //     payload: { popular: res.data.items },
+    //   });
+    // });
   }, []);
 
   return (
